requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
pydantic==2.11.9
transformers== 4.56.2
torch==2.8.0
python-dateutil==2.9.0.post0
cachetools==5.5.2
sentence-transformers==5.1.1
faiss-cpu==1.12.0
python-dotenv==1.1.0
sacremoses==0.1.1
sentencepiece== 0.2.1